requirements.txt
wmi
face_recognition==1.3.0
tensorflow_gpu==2.1.0
matplotlib==3.3.4
numpy==1.19.5
mttkinter==0.6.1
Pillow==8.2.0
scikit_learn==0.24.2
tensorflow==2.5.0
